split FATHMM pathogenicity into BP4/PP3
FATHMM, unlike the other calibrated variant effect predictors,
currently uses a single reporting column that puts both the BP4 and
PP3 into a single result column. However in the reporting widgets,
there's a single column for both.

This commit splits the annotator output into two columns, and then
rejoins the columns in the web viewer widget, to match the behavior of
the other calibrations.